Add low-level parseSegment function for routing
This started with my need to parse something with an Enum instance, so
it was approximately an Int but needed a way to fail. Pux does not
export constructors for Match (which seems sensible), but the
typeclasses do not seem to provide a way to depend on the result of a
match and then conditionally fail. This could probably be achieved
with Monad but I'm not sure if that's necessary. Instead I provided a
low level parser that may fail and reimplemented the existing parser
in terms of it.